% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_sequence.R
\name{draw_sequence}
\alias{draw_sequence}
\title{draw a sequence, for examples}
\usage{
draw_sequence(
  state_seq,
  x = 1:length(state_seq) - 1,
  labels,
  cols,
  y = 0,
  box = TRUE,
  ...
)
}
\arguments{
\item{state_seq}{character vector with state names in each time step}

\item{x}{left bound of each time step, possibly after alignment}

\item{labels}{vector of time step labels, potentially after clocking}

\item{cols}{vector of colors, with names equal to state space.}

\item{y}{height at which to draw the sequence, assumed 1 unit tall.}

\item{box}{logical do we draw a box outline around the whole sequence?}

\item{...}{optional arguments passed to \code{rect()} or \code{text()}}
}
\description{
draws a single sequence using base functions. For use in illustrative diagrams
}
\details{
if you want to indicate a cesored spell, then you need to include a color for that as well. Since a seqence or a spell can be censored, it may as well be a nice thing to add to the color palette in general, so that it kicks in faithefully whether for individual time steps or for the full sequence. 

We assume the trajectory height is equal to 1, and also that time steps are of width 1.

If \code{box}is \code{TRUE} then any state not equal to \code{"Dead"} is included in the box. Non-flexible aesthetics, sorry.
}
\examples{
\dontrun{
library(Spells)
data(X)
library(colorspace)
cols  <- c(qualitative_hcl(5, palette = "Dark 3")[c(3,2,4)],NA)
cols2 <- lighten(desaturate(cols,.3),.3)
names(cols) <- names(cols2) <- c("Employed", "Inactive","Retired","Dead")

plot(NULL, 
     type = "n", 
     xlim = c(-50,0), 
     ylim = c(0,12), 
     axes = FALSE, 
     xlab = "", 
     ylab = "")
for (i in 1:10){
  draw_sequence(X[,i], 
                x = align(X[,i], 
                          state = "Dead",
                          type = "left"),
                labels = clock(X[,i], 
                               state = "Retired", 
                               clock_type = "step", 
                               increasing = FALSE),
                cols = cols2,
                box = TRUE,
                y = i,
                border = NA,
                cex=.5
                )
}

}
}
