% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions.R
\name{intervalpoly}
\alias{intervalpoly}
\title{draw a polygon, used for interval fans}
\usage{
intervalpoly(x, u, l, ...)
}
\arguments{
\item{x}{numeric vector of time steps}

\item{u}{numeric vector of upper bound}

\item{l}{numeric vector of lower bound}

\item{...}{optional arguments to pass to \code{polygon()}}
}
\value{
function renders a polygon in the graphics device.
}
\description{
Draw a confidence band
}
\details{
upper and lower bounds must have the same time steps in \code{x}
}
