% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions.R
\name{qdens}
\alias{qdens}
\title{estimate quantiles from a matrix of sequences}
\usage{
qdens(X, p = 0.5)
}
\arguments{
\item{X}{numeric matrix, with time steps in rows and individuals in columns}

\item{p}{desired quantiles}
}
\value{
numeric vector (matrix) of the desired quantile(s) per time step.
}
\description{
Within a time step, first fit a smooth density, then estimate a given quantile on the density.
}
\details{
if \code{p} is given as a vector, then the result is a matrix with percentiles in rows and time steps in columns.
}
\examples{
X <- runif(5000)
dim(X) <- c(50,100)
qdens(X, p = .5)
qdens(X, p = c(.25, .75))
}
