% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_sequence.R
\docType{data}
\name{X}
\alias{X}
\title{example trajectory data}
\format{
A character matrix, with individual trajectories in columns and age in rows (ages 50-100)
}
\source{
Simulated based on transition probabilities published in:
  Dudel, Christian, and Mikko Myrskylä. "Working life expectancy at age 50 in the United States and the impact of the great recession." Demography 54.6 (2017): 2101-2123.
}
\usage{
X
}
\description{
Ten simulated labor trajectories, used in examples.
}
\keyword{datasets}
