% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions.R
\name{intervalfan}
\alias{intervalfan}
\title{draw a polygon, used for interval fans}
\usage{
intervalfan(X, x = 50:100, alphas = seq(0.1, 0.9, by = 0.1), ...)
}
\arguments{
\item{x}{numeric vector of time steps}

\item{...}{optional arguments to pass to \code{polygon()}}

\item{u}{numeric vector of upper bound}

\item{l}{numeric vector of lower bound}
}
\value{
function renders a polygon in the graphics device.
}
\description{
Draw a confidence band
}
\details{
upper and lower bounds must have the same time steps in \code{x}
}
\examples{
X <- runif(5000)
dim(X) <- c(50,100)
plot(NULL, type="n",xlim=c(0,50),ylim = c(0,1))
intervalfan(X,x=0:49,alphas=seq(.1, .9, by = .1),col = "#FF000020")
}
